+++ CHANGELOG.md
@@ -3,9 +3,19 @@
 <a name="unreleased"></a>
 ## [Unreleased]
 
+### New Features
+- use SQLAlchemy for the  Rekordbox6 `master.db` database
+
 ### Improvements/Bug Fixes
+- fix import error and README.md
+- set logging level to warning
+- fix loading the Rekordbox setting file twice in config initialization
+- add context for Rekordbox 6 database
 - inherit AnlzFile from Mapping to implement dict interface
 - unify binary file API
+
+### Documentation
+- add missing djmd tables to `master.db` database documentation
 
 
 <a name="0.0.1"></a>